Rabbitmq reported not running in zabbix
RabbitMQ is reported as not running when rabbitmq is run
as root user. This patch removes the check as which
user rabbitmq is running.

This may introduce a security issue if our purpose was to
make sure that rabbitmq is always run as rabbitmq user.
If so please discard this patch and change init script
of non-ha mode to run RabbitMQ as rabbitmq user.

The fuel-plugin-external-zabbix did not backport latest
code of Fuel, causing a regression. The following
patch was not ported to fuel plugin:
https://review.openstack.org/153840

Change-Id: Id9106d178a300880e6e491e59cb3bc3245b8bded
Closes-bug: #1419397